1 Part, version 1.1

+++ src/main/java/com/example/website/repos/MessageRepo.java
@@ -1,9 +1,9 @@
 package com.example.website.repos;
 
-import com.example.website.Message;
+import com.example.website.domain.Message;
 import org.springframework.data.repository.CrudRepository;
 import java.util.List;
 
-public interface MessageRepo extends CrudRepository<Message, Integer> {
+public interface MessageRepo extends CrudRepository<Message, Long> {
     List<Message> findByTag(String tag);
 }
